MultiAgentPOMDPs

MultiAgentPOMDPs.agent_actionindex

MultiAgentPOMDPs.agent_actions

MultiAgentPOMDPs.agent_stateindex

MultiAgentPOMDPs.agent_states

MultiAgentPOMDPs.coordination_graph

MultiAgentPOMDPs.n_agents

MultiAgentPOMDPs.agent_actionindex — Function

function agent_actionindex(m::JointMDP, idx::Int64, a::A) where A

Returns the integer index of action a for agent idx. Used for discrete models only.

MultiAgentPOMDPs.agent_actions — Function

function agent_actions(m::JointMDP, idx::Int64, s::S) where S

Returns the discrete actions for the given agent index.

Note

This will be called a LOT so it should not allocate each time....

MultiAgentPOMDPs.agent_stateindex — Function

function agent_actionindex(m::JointMDP, idx::Int64, s::S) where S

Returns the integer index of state s for agent idx. Used for discrete models only.

MultiAgentPOMDPs.agent_states — Function

function agent_states(m::JointMDP, idx::Int64)

Returns the discrete states for the given agent index

MultiAgentPOMDPs.coordination_graph — Function

function coordination_graph(m::JointMDP)
function coordination_graph(m::JointMDP, s::S) where S

Returns the LightGraphs.SimpleGraph (or any appropriate structure) for the coordination graph.

MultiAgentPOMDPs.n_agents — Function

function n_agents(m::JointMDP)

Return the number of agents in the (PO)MDP